Additional changes after code review.
Added a check to remove potential padding bits at the end of the bit vector.

Added some test bit vectors (they are also used for unit tests in SC Crypto Lib).

Switched to the "development_tmp" branch of the Ginger Lib in order to use the updated merkle tree parameters.